ops-agent/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-google-genai  # <-- ADDED BACK
faiss-cpu
python-dotenv
pydantic
unstructured
markdown
slack-sdk